[style] 회원가입 약관동의 페이지 UI #95
